% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geodata.df.R
\docType{data}
\name{geodata.df}
\alias{geodata.df}
\title{Yeast strain geographic metadata (preloaded)}
\format{
A tibble with columns:
\describe{
\item{std_name}{Strain identifier (character)}
\item{StrainName}{Original strain name (character)}
\item{Reference}{Reference/source (factor)}
\item{SRA}{SRA accession (character)}
\item{Clade}{Genetic clade (character)}
\item{geographic_value}{Geographical descriptor (character)}
\item{city_region_value}{City or region (character)}
\item{province_state_value}{Province/state (character)}
\item{country_value}{Country (character)}
\item{geo_cc}{Geographic country code (factor)}
\item{geo_country}{Geographic country name (factor)}
\item{geo_region}{Geographic region (factor)}
\item{geo_lat}{Latitude (numeric)}
\item{geo_lon}{Longitude (numeric)}
\item{geo_label}{Geographic label (factor)}
}
}
\usage{
geodata.df
}
\description{
Contains geographic coordinates and metadata for yeast strains.
}
\seealso{
\code{\link{read_geodata.df}}
}
\keyword{datasets}
