% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_yeast_virome}
\alias{load_yeast_virome}
\title{Load Yeast Virome CSV Data}
\usage{
load_yeast_virome()
}
\value{
A named \code{list} of \code{tibble}s.
}
\description{
Reads every \code{.csv} file in \verb{inst/extdata/} and returns a named list
of tibbles.  File names become list names.
}
\examples{
data_list <- load_yeast_virome()
names(data_list)
}
