% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_killerdata.df.R
\name{read_killerdata.df}
\alias{read_killerdata.df}
\title{Read yeast strain killer data}
\usage{
read_killerdata.df(
  path = system.file("extdata", "strain_killerdata.csv", package = "yeastvirome")
)
}
\arguments{
\item{path}{Optional path override.}
}
\value{
A tibble with columns:
\describe{
\item{std_name}{Strain identifier (character)}
\item{Collection}{Collection from which strain is sourced (factor)}
\item{Killer}{Strain is a killer (logical)}
\item{Only_Y_2046}{Strain is not a killer, kills hypersusceptible strain Y-2046 only (logical)}
\item{Y_2046}{Strain kills Y-2046 (logical)}
\item{Y_1088}{Strain kills Y-1088 (logical)}
\item{DSM70459}{Strain kills DSM70459 (logical)}
\item{CYC1058}{Strain kills CYC1058 (logical)}
\item{Y-27106}{Strain kills Y-27106 (logical)}
\item{NCYC1006}{Strain kills NCYC1006 (logical)}
\item{NCYC777}{Strain kills NCYC777 (logical)}
\item{NBRC1802}{Strain kills NBRC1802 (logical)}
\item{NBRC815}{Strain kills NBRC815 (logical)}
\item{Y_5509}{Strain kills Y-5509 (logical)}
\item{MS300c}{Strain kills MS300c (logical)}
\item{BY4741}{Strain kills BY4741 (logical)}
\item{KHR}{KHR1 gene status (factor)}
\item{KHS}{KHS1 gene status (factor)}
\item{Func_KHR}{Functional KHR protein (logical)}
\item{Func_KHS}{Functional KHS protein (logical)}
}
}
\description{
Loads strain killer activity data from \strong{strain_killerdata.csv}, containing
strain identifiers, killer activity with respect to 12 test strains, and KHR
and KHS status
}
