% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rnaseq_accessions.R
\name{read_rnaseq_accessions}
\alias{read_rnaseq_accessions}
\title{Read RNA-seq accession metadata}
\usage{
read_rnaseq_accessions(
  path = system.file("extdata", "rnaseq_accessions.tsv", package = "yeastvirome")
)
}
\arguments{
\item{path}{Optional override path.}
}
\value{
A tibble with columns:
\describe{
\item{tree_order}{Integer tree order for plotting/ordering.}
\item{std_name}{Yeast strain name (character).}
\item{accession}{ENA accession ID (character).}
\item{replicate}{Replicate number (integer).}
}
}
\description{
Loads \strong{inst/extdata/rnaseq_accessions.tsv}, containing metadata for
RNA-seq runs in the yeast virome project.
}
