% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_virome.df.R
\name{read_virome.df}
\alias{read_virome.df}
\title{Read wide-format contig abundance matrix}
\usage{
read_virome.df(
  path = system.file("extdata", "virome250531.tsv", package = "yeastvirome")
)
}
\arguments{
\item{path}{Optional path override. Defaults to internal data.}
}
\value{
A wide-format tibble with columns:
\describe{
\item{accession}{Strain name (character).}
\item{\if{html}{\out{<contig ids>}}}{Numeric abundance values (e.g. RPM).}
}
}
\description{
Loads \strong{inst/extdata/virome250531.tsv}, which contains virome abundance
data for each yeast strain. Each row corresponds to a yeast strain and each
column (after \code{accession}) corresponds to a viral contig.
}
