% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_virome.R
\name{summarize_virome_family}
\alias{summarize_virome_family}
\title{Summarize Virome Data by Family}
\usage{
summarize_virome_family(virome.df, vmeta.df)
}
\arguments{
\item{virome.df}{The original wide data.frame (samples x viruses)}

\item{vmeta.df}{The metadata data.frame mapping viruses to species and family}
}
\value{
Family-level summary data.frame
}
\description{
Summarize Virome Data by Family
}
