% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_virome.R
\name{summarize_virome_species}
\alias{summarize_virome_species}
\title{Summarize Virome Data by Species}
\usage{
summarize_virome_species(virome.df, virometa.df)
}
\arguments{
\item{virome.df}{The original wide data.frame (samples x viruses)}

\item{virometa.df}{The metadata data.frame mapping viruses to species and family}
}
\value{
Species-level summary data.frame
}
\description{
Summarize Virome Data by Species
}
