% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_virome.R
\name{summarize_virome_total}
\alias{summarize_virome_total}
\title{Summarize Total Virome Data}
\usage{
summarize_virome_total(virome.df, virometa.df)
}
\arguments{
\item{virome.df}{The original wide data.frame (samples x viruses)}

\item{virometa.df}{The metadata data.frame mapping viruses to species and family}
}
\value{
Total summary data.frame
}
\description{
Summarize Total Virome Data
}
