% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_virometa.df.R
\docType{data}
\name{virometa.df}
\alias{virometa.df}
\title{Viral contig taxonomy metadata (preloaded)}
\format{
A tibble with the columns:
\describe{
\item{rvlabel}{Contig label (character)}
\item{accession}{Yeast strain accession (character)}
\item{species}{Viral species classification (character)}
\item{family}{Viral family classification (character)}
}
}
\usage{
virometa.df
}
\description{
Preloaded tibble with viral contig metadata, including strain accession and
taxonomy (species and family).
}
\seealso{
\code{\link{read_virometa.df}}
}
\keyword{datasets}
