% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeastvirome-package.R
\docType{package}
\name{yeastvirome}
\alias{yeastvirome-package}
\alias{yeastvirome}
\title{yeastvirome: Explore the 1011 Yeast RNA Virome}
\description{
Core package documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ababaian/yeastvirome}
  \item Report bugs at \url{https://github.com/ababaian/yeastvirome/issues}
}

}
\author{
\strong{Maintainer}: Artem Babaian \email{artem@rRNA.ca}

}
\keyword{internal}
